move translated layer
