+++ python/paddle/fluid/dygraph/__init__.py
@@ -36,10 +36,6 @@ from .checkpoint import *
 from . import learning_rate_scheduler
 from .learning_rate_scheduler import *
 
-
-from . import io
-from .io import *
-
 from . import static_runner
 from .static_runner import StaticModelRunner
