+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_lac.py
@@ -26,10 +26,10 @@ import paddle
 import paddle.fluid as fluid
 from paddle import _legacy_C_ops
 from paddle.fluid.dygraph import to_variable
-from paddle.fluid.dygraph.io import INFER_MODEL_SUFFIX, INFER_PARAMS_SUFFIX
 from paddle.fluid.framework import _non_static_mode
 from paddle.jit import ProgramTranslator
 from paddle.jit.api import declarative
+from paddle.jit.translated_layer import INFER_MODEL_SUFFIX, INFER_PARAMS_SUFFIX
 
 SEED = 2020
