+++ python/paddle/jit/__init__.py
@@ -17,11 +17,11 @@ from .api import save
 from .api import load
 from .api import declarative as to_static
 from .api import not_to_static
-from .api import TranslatedLayer
 from .dy2static.logging_utils import set_code_level, set_verbosity
 
 from . import dy2static
 from .dy2static.program_translator import ProgramTranslator
+from .translated_layer import TranslatedLayer
 
 __all__ = [  # noqa
     'save',
