+++ python/paddle/jit/dy2static/function_spec.py
@@ -21,8 +21,8 @@ import paddle
 from paddle.fluid import core
 from paddle.fluid.dygraph import layers
 from paddle.fluid.dygraph.base import switch_to_static_graph
-from paddle.fluid.dygraph.io import TranslatedLayer
 from paddle.fluid.layers.utils import flatten, pack_sequence_as
+from paddle.jit.translated_layer import TranslatedLayer
 
 from . import logging_utils
 from .utils import (
